+++ src/helpers/mod.rs
@@ -1,0 +1,3 @@
+mod get_subdirs;
+
+pub use get_subdirs::get_subdirs;
