+++ src/main.rs
@@ -2,6 +2,7 @@ mod cli;
 mod commands;
 mod error;
 mod github;
+mod helpers;
 mod operations;
 mod runtime;
 
@@ -28,17 +29,14 @@ fn inner() -> Result<(), UserError> {
     let args = cli::Arguments::parse();
     let persisted_steps = runtime::load()?;
     let initial_dir = env::current_dir().expect("cannot determine the current directory");
+    let initial_dir = Utf8PathBuf::from_path_buf(initial_dir).expect("invalid unicode in filename");
     let current_steps = match args.command {
         Command::Abort => commands::abort(&persisted_steps)?,
         Command::Clone { org } => commands::clone(&org),
-        Command::Run { cmd, args } => {
-            let file =
-                Utf8PathBuf::from_path_buf(initial_dir).expect("invalid unicode in filename");
-            commands::run(&cmd, &args, file)?
-        }
+        Command::Run { cmd, args } => commands::run(&cmd, &args, initial_dir)?,
         Command::Ignore => commands::ignore(persisted_steps)?,
         Command::Retry => commands::retry(persisted_steps)?,
-        Command::Walk => commands::walk(),
+        Command::Walk => commands::walk(initial_dir)?,
     };
     match runtime::execute(current_steps) {
         Outcome::Success => {
