+++ src/operations/exit.rs
@@ -1,3 +1,5 @@
+use crate::runtime::Step;
+
 pub fn exit(id: u32) -> Step {
-    Step::Exit{id}
+    Step::Exit { id }
 }
