+++ src/operations/mod.rs
@@ -1,7 +1,9 @@
 mod chdir;
 mod clone_repo;
 mod execute;
+mod exit;
 
 pub use chdir::chdir;
 pub use clone_repo::clone_repo;
 pub use execute::execute;
+pub use exit::exit;
